// src/app/dto/mapAction.enum.ts
export enum MapAction {
    GET_ENVELOPPE = 'GET_ENVELOPPE',
    GET_POTENTIAL = 'GET_POTENTIAL',
    EDIT_ENVELOPPE = 'EDIT_ENVELOPPE',
    EDIT_POTENTIEL = 'EDIT_POTENTIEL',
    CANCEL_EDIT_ENVELOPPE = 'CANCEL_EDIT_ENVELOPPE',
    CANCEL_EDIT_POTENTIEL = 'CANCEL_EDIT_POTENTIEL',
    SAVE_ENVELOPPE = 'SAVE_ENVELOPPE',
    SAVE_POTENTIEL = 'SAVE_POTENTIEL',
    REMOVE_ENVELOPPE = 'REMOVE_ENVELOPPE',
    REMOVE_POTENTIEL = 'REMOVE_POTENTIEL',
    FLY_TO = "FLY_TO",
    
    NO_PROCESS_DATA = 'NO_PROCESS_DATA'
}